shell

+++ main.py
@@ -29,16 +29,16 @@ def main():
     parser.add_argument('-outfile')
     parser.add_argument('-mode', required = True)
     args = parser.parse_args()
-    size = args.size if args.size != None else 256
+    size = int(args.size) if args.size != None else 256
     augment = True if args.augment != None else False
     checkpoint = args.checkpoint if args.checkpoint != None else './Model/Core.ckpt'
-    lr = args.lr if args.lr != None else 0.0001
-    dlr = args.dlr if args.dlr != None else 0.0
+    lr = float(args.lr) if args.lr != None else 0.0001
+    dlr = float(args.dlr) if args.dlr != None else 0.0
     rtype = args.rtype if args.rtype != None else 'L2'
     rr = args.rr if args.rr != None else 0.0
-    epoch = args.epoch if args.epoch != None else 2
-    batch = args.batch if args.batch != None else 5
-    log = args.log if args.log != None else 2
+    epoch = int(args.epoch) if args.epoch != None else 2
+    batch = int(args.batch) if args.batch != None else 5
+    log = int(args.log) if args.log != None else 2
     outfile = args.outfile if args.outfile != None else './Result/result.csv'
     if args.mode in ['learning']:
         init = True
